+++ packages/compiler-cli/src/typescript_support.ts
@@ -7,7 +7,7 @@
  */
 import ts from 'typescript';
 
-import {compareVersions} from './diagnostics/typescript_version';
+import {compareVersions} from './version_helpers';
 
 /**
  * Minimum supported TypeScript version
