--- packages/compiler-cli/test/diagnostics/BUILD.bazel
@@ -1,20 +1,0 @@
-load("//tools:defaults.bzl", "jasmine_node_test", "ts_library")
-
-# typescript_version_spec
-ts_library(
-    name = "typescript_version_lib",
-    testonly = True,
-    srcs = ["typescript_version_spec.ts"],
-    deps = [
-        "//packages/compiler-cli",
-        "//packages/compiler-cli/test:test_utils",
-    ],
-)
-
-jasmine_node_test(
-    name = "typescript_version",
-    bootstrap = ["//tools/testing:node_es5"],
-    deps = [
-        ":typescript_version_lib",
-    ],
-)
